make rails 2 compatible
This code was recently added and is not rails 2 compatible because it
assumes the context of the block is the instance and that it can call
the `:algolia_synchronous?` private method on it. In rails 2, the
context of the hook was the class itself, so this fails. To be
backwards-compatible, we can use `instance_exec` to simulate the newer
rails behavior.